Moved app to separate package
