+++ sciencebeam_parser/service/server.py
@@ -9,7 +9,7 @@ import uvicorn
 
 from sciencebeam_parser.app.parser import ScienceBeamParser
 from sciencebeam_parser.config.config import AppConfig
-from sciencebeam_parser.service.routers.api import create_api_app
+from sciencebeam_parser.service.api.app import create_api_app
 from sciencebeam_parser.service.routers.index import create_index_router
 from sciencebeam_parser.resources.default_config import DEFAULT_CONFIG_FILE
